Transition to server-based code
Upon program execution, automatically creates a lobby and first user.  Fixed add/remove user functions so that server and chatroom lists are in synch. (Ex: if a user is removed from a chatroom and back to the lobby, the numbers increment correctly. If a user is booted completely from the server due to admin priviledge, he is deleted from chatroom.